created posts controllers
